+++ NEWS.md
@@ -1,10 +1,12 @@
-# emuR 1.9.9.9001
+# emuR 1.9.9.9002
 
 ## new features / performance tweaks / improvements
 
 * changed default `resultTypes` to `"tibble"` (`query()`, `requery_hier()`, `requery_seq()`, `get_trackdata()`)
 * changed url in printed output of serve to https
 * `serve()` doesn't block the R console any more
+* `load_emuDB()` can now automatically create a load snapshot of an emuDB using git (using the git2r package). For this, set
+the new `gitSnapshot` parameter to `TRUE`.
 * suppressing requery differing length warnings in `runBASwebservice_*` functions
 * `requery_hier()` now persists the input segment list length by inserting NA rows for missing segments
 * depricated vignettes are now stubs only (== empty containing notice pointing to the EMU-SDMS manual)